requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
django-environ==0.11.2
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1